Create README.md
Thread 1: creates 30 new patients and puts them in the InLine Table using entity framework.

Thread 2 : Sleeps for 5 seconds then picks up the patients from the Inline table and puts them in IVA or Sanatorium. Where they get stored depends on there age and sickness level. Only 5 patients are allowed in IVA and 10 in the sanatorium. If there is a free spot in IVA the sickesst patient should be moved there, no matter if they are Inline or at the sanatorium table.

Thread 3 Sleeps for 3 seconds, the change the symtom level with a serten set of rules. These can be found in the GeneratePercentage class.

Thread 4 Sleeps for 5 seconds, then moves the patients from IVA and Sanatorium to Recovery or AfterLife.